refac: Move `csv_column_names` from infra to domain
