+++ source/databricks/settlement_report/settlement_report_job/domain/energy_results_factory.py
@@ -20,7 +20,7 @@ import settlement_report_job.domain.report_naming_convention as market_naming
 from settlement_report_job.domain.repository import WholesaleRepository
 from settlement_report_job.domain.settlement_report_args import SettlementReportArgs
 
-from settlement_report_job.infrastructure.csv_column_names import (
+from settlement_report_job.domain.csv_column_names import (
     EnergyResultsCsvColumnNames,
 )
 from settlement_report_job.utils import map_from_dict
