+++ source/databricks/settlement_report/tests/test_factories/time_series_csv_factory.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 
 from pyspark.sql import SparkSession, DataFrame
 
-from settlement_report_job.infrastructure.csv_column_names import (
+from settlement_report_job.domain.csv_column_names import (
     TimeSeriesPointCsvColumnNames,
 )
 from settlement_report_job.wholesale.column_names import DataProductColumnNames
